+++ pcmc/bdd/enums.py
@@ -1,0 +1,24 @@
+"""pcmc-bot / bdd / Enums
+
+Déclaration des énumérations utilisées
+
+"""
+
+import enum
+
+
+# Définition des enums
+
+# class Statut(enum.Enum):
+#     """:class:`~enum.Enum` représentant le statut RP d'un Joueur.
+#
+#     Attributes:
+#         vivant: Le joueur est en vie !
+#         mort: Le joueur est mort. RIP.
+#         MV: Le joueur est Mort-Vivant. Pas de chance.
+#         immortel: Le joueur est Immortel. Si jamais...
+#     """
+#     vivant = enum.auto()
+#     mort = enum.auto()
+#     MV = enum.auto()
+#     immortel = enum.auto()